rsma1-bk-nopc-new-case-entry/requirements.txt
rpaframework==10.*
PyPDF2
python-docx
pytesseract
pdf2image
PyMuPDF
ta_bitwarden_cli
